moved files around for both pop ups to have access, updated snaps
