+++ src/app/content/popUpElements/ChapterFilter.spec.tsx
@@ -2,18 +2,18 @@ import { HighlightColorEnum } from '@openstax/highlighter/dist/api';
 import React from 'react';
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import createTestStore from '../../../../../test/createTestStore';
-import { book as archiveBook, page } from '../../../../../test/mocks/archiveLoader';
-import { mockCmsBook } from '../../../../../test/mocks/osWebLoader';
-import AllOrNone from '../../../../components/AllOrNone';
-import Checkbox from '../../../../components/Checkbox';
-import MessageProvider from '../../../../MessageProvider';
-import { Store } from '../../../../types';
-import { assertDefined } from '../../../../utils';
-import { receiveBook, receivePage } from '../../../actions';
-import { formatBookData } from '../../../utils';
-import { findArchiveTreeNode } from '../../../utils/archiveTreeUtils';
-import { receiveHighlightsTotalCounts } from '../../actions';
+import createTestStore from '../../../test/createTestStore';
+import { book as archiveBook, page } from '../../../test/mocks/archiveLoader';
+import { mockCmsBook } from '../../../test/mocks/osWebLoader';
+import AllOrNone from '../../components/AllOrNone';
+import Checkbox from '../../components/Checkbox';
+import MessageProvider from '../../MessageProvider';
+import { Store } from '../../types';
+import { assertDefined } from '../../utils';
+import { receiveBook, receivePage } from '../actions';
+import { receiveHighlightsTotalCounts } from '../highlights/actions';
+import { formatBookData } from '../utils';
+import { findArchiveTreeNode } from '../utils/archiveTreeUtils';
 import ChapterFilter from './ChapterFilter';
 
 describe('ChapterFilter', () => {
