+++ src/app/content/popUpElements/ColorFilter.tsx
@@ -3,16 +3,16 @@ import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components/macro';
-import AllOrNone from '../../../../components/AllOrNone';
-import Checkbox from '../../../../components/Checkbox';
-import { textStyle } from '../../../../components/Typography/base';
-import { match, not } from '../../../../fpUtils';
-import theme from '../../../../theme';
-import { highlightStyles } from '../../../constants';
-import { filters } from '../../../styles/PopupConstants';
-import { setSummaryFilters } from '../../actions';
-import { highlightColorFiltersWithContent, summaryColorFilters } from '../../selectors';
-import ColorIndicator from '../ColorIndicator';
+import AllOrNone from '../../components/AllOrNone';
+import Checkbox from '../../components/Checkbox';
+import { textStyle } from '../../components/Typography/base';
+import { match, not } from '../../fpUtils';
+import theme from '../../theme';
+import { highlightStyles } from '../constants';
+import { setSummaryFilters } from '../highlights/actions';
+import ColorIndicator from '../highlights/components/ColorIndicator';
+import { highlightColorFiltersWithContent, summaryColorFilters } from '../highlights/selectors';
+import { filters } from '../styles/PopupConstants';
 
 interface Props {
   className?: string;
