+++ src/app/content/popUpElements/Filters.spec.tsx
@@ -2,20 +2,20 @@ import { HighlightColorEnum } from '@openstax/highlighter/dist/api';
 import React from 'react';
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import createTestServices from '../../../../../test/createTestServices';
-import createTestStore from '../../../../../test/createTestStore';
-import { book as archiveBook, pageInChapter } from '../../../../../test/mocks/archiveLoader';
-import { mockCmsBook } from '../../../../../test/mocks/osWebLoader';
-import { DropdownToggle } from '../../../../components/Dropdown';
-import MessageProvider from '../../../../MessageProvider';
-import { MiddlewareAPI, Store } from '../../../../types';
-import { assertDefined } from '../../../../utils';
-import { receiveBook, receivePage } from '../../../actions';
-import { ArchiveTree } from '../../../types';
-import { formatBookData } from '../../../utils';
-import { findArchiveTreeNode } from '../../../utils/archiveTreeUtils';
-import { stripIdVersion } from '../../../utils/idUtils';
-import { receiveHighlightsTotalCounts, setSummaryFilters } from '../../actions';
+import createTestServices from '../../../test/createTestServices';
+import createTestStore from '../../../test/createTestStore';
+import { book as archiveBook, pageInChapter } from '../../../test/mocks/archiveLoader';
+import { mockCmsBook } from '../../../test/mocks/osWebLoader';
+import { DropdownToggle } from '../../components/Dropdown';
+import MessageProvider from '../../MessageProvider';
+import { MiddlewareAPI, Store } from '../../types';
+import { assertDefined } from '../../utils';
+import { receiveBook, receivePage } from '../actions';
+import { receiveHighlightsTotalCounts, setSummaryFilters } from '../highlights/actions';
+import { ArchiveTree } from '../types';
+import { formatBookData } from '../utils';
+import { findArchiveTreeNode } from '../utils/archiveTreeUtils';
+import { stripIdVersion } from '../utils/idUtils';
 import Filters from './Filters';
 import { FiltersListChapter, FiltersListColor, StyledPlainButton } from './FiltersList';
